Move join form to component, add hooks
